Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main(){
-    int radius,area;
+    float radius,area;
     float PI=3.14;
-    scanf("%d",&radius);
+    scanf("%f",&radius);
     area= PI*radius*radius;
     printf("Area: %.2f",area);
     return 0;
